refactor: rename browser-protocol to peersky-protocol for clarity
