+++ src/main.js
@@ -1,5 +1,5 @@
 import { app, session, protocol as globalProtocol } from "electron";
-import { createHandler as createBrowserHandler } from "./protocols/browser-protocol.js";
+import { createHandler as createBrowserHandler } from "./protocols/peersky-protocol.js";
 import { createHandler as createBrowserThemeHandler } from "./protocols/theme-handler.js";
 import { createHandler as createIPFSHandler } from "./protocols/ipfs-handler.js";
 import { createHandler as createHyperHandler } from "./protocols/hyper-handler.js";
